Update ch04-02-references-and-borrowing.md
Explain dangling reference prevention better.

-------------

The original wording might lead someone learning Rust to think that data will be kept around for as long as a reference exists. Instead, as soon as data is dropped, the reference is barred from being used. Those are two different mechanisms, and this disambiguates which one is in place.